requirements.txt
openai
gitpython
python-dotenv
transformers
torch
